backend/requirements.txt
Flask==1.1.1
PyMongo==3.10.0
requests==2.20.1
jsonify==0.5
python-dotenv==0.10.3
connexion==2.5.1
swagger-ui-bundle==0.0.6
pytest==5.3.2
matplotlib==3.1.2
flask-cors==3.0.8
